add genesis module, add parsing in the beacon node
