+++ lib/lambda_ethereum_consensus/application.ex
@@ -44,15 +44,10 @@ defmodule LambdaEthereumConsensus.Application do
   defp get_children(:full) do
     get_children(:db) ++
       [
-        {LambdaEthereumConsensus.Beacon.BeaconNode, [checkpoint_sync_url()]},
+        LambdaEthereumConsensus.Beacon.BeaconNode,
         LambdaEthereumConsensus.P2P.Metadata,
         BeaconApi.Endpoint
       ]
-  end
-
-  def checkpoint_sync_url do
-    Application.fetch_env!(:lambda_ethereum_consensus, LambdaEthereumConsensus.ForkChoice)
-    |> Keyword.fetch!(:checkpoint_sync_url)
   end
 
   defp get_operation_mode do
