+++ lib/lambda_ethereum_consensus/beacon/beacon_node.ex
@@ -19,32 +19,12 @@ defmodule LambdaEthereumConsensus.Beacon.BeaconNode do
   end
 
   @impl true
-  def init([nil]) do
-    case restore_state_from_db() do
-      nil ->
-        Logger.error(
-          "[Sync] No recent state found. Please specify the URL to fetch them from via the --checkpoint-sync-url flag"
-        )
+  def init(_) do
+    {store, genesis_validators_root} =
+      Application.get_env(:lambda_ethereum_consensus, ForkChoice)
+      |> Keyword.fetch!(:genesis)
+      |> Genesis.get_state!()
 
-        System.halt(1)
-
-      {_, {store, root}} ->
-        init_children(store, root)
-    end
-  end
-
-  def init([checkpoint_url]) do
-    case restore_state_from_db() do
-      {:ok, {store, root}} ->
-        Logger.warning("[Checkpoint sync] Recent state found. Ignoring the checkpoint URL.")
-        init_children(store, root)
-
-      _ ->
-        fetch_state_from_url(checkpoint_url)
-    end
-  end
-
-  defp init_children(%Store{} = store, genesis_validators_root) do
     Cache.initialize_cache()
 
     config = Application.fetch_env!(:lambda_ethereum_consensus, :discovery)
@@ -86,72 +66,5 @@ defmodule LambdaEthereumConsensus.Beacon.BeaconNode do
     ]
 
     Supervisor.init(children, strategy: :one_for_all)
-  end
-
-  defp restore_state_from_db do
-    # Try to fetch the old store from the database
-    case StoreDb.fetch_store() do
-      {:ok, %Store{finalized_checkpoint: %{epoch: finalized_epoch}} = store} ->
-        res = {store, ChainSpec.get_genesis_validators_root()}
-
-        if get_current_epoch(store) - finalized_epoch > @max_epochs_before_stale do
-          Logger.info("[Sync] Found old state in DB.")
-          {:old_state, res}
-        else
-          Logger.info("[Sync] Found recent state in DB.")
-          {:ok, res}
-        end
-
-      :not_found ->
-        nil
-    end
-  end
-
-  defp fetch_state_from_url(url) do
-    Logger.info("[Checkpoint sync] Initiating checkpoint sync")
-
-    genesis_validators_root = ChainSpec.get_genesis_validators_root()
-
-    case CheckpointSync.get_finalized_block_and_state(url, genesis_validators_root) do
-      {:ok, {anchor_state, anchor_block}} ->
-        Logger.info(
-          "[Checkpoint sync] Received beacon state and block",
-          slot: anchor_state.slot
-        )
-
-        # We already checked block and state match
-        {:ok, store} = Store.get_forkchoice_store(anchor_state, anchor_block)
-
-        # TODO: integrate into CheckpointSync, and validate snapshot
-        snapshot = fetch_deposit_snapshot(url)
-        store = Store.init_deposit_tree(store, snapshot)
-
-        # Save store in DB
-        StoreDb.persist_store(store)
-
-        init_children(store, genesis_validators_root)
-
-      _ ->
-        Logger.error("[Checkpoint sync] Failed to fetch the latest finalized state and block")
-
-        System.halt(1)
-    end
-  end
-
-  defp get_current_epoch(store) do
-    (:os.system_time(:second) - store.genesis_time)
-    |> div(ChainSpec.get("SECONDS_PER_SLOT"))
-    |> Misc.compute_epoch_at_slot()
-  end
-
-  defp fetch_deposit_snapshot(url) do
-    case CheckpointSync.get_deposit_snapshot(url) do
-      {:ok, snapshot} ->
-        snapshot
-
-      _ ->
-        Logger.error("[Checkpoint sync] Failed to fetch the deposit snapshot")
-        System.halt(1)
-    end
   end
 end
